refactor #28 : Product 관련 DTO 리팩토링
