+++ src/main/java/com/antique/controller/user/UserDibsController.java
@@ -1,6 +1,6 @@
 package com.antique.controller.user;
 
-import com.antique.dto.ProductDTO;
+import com.antique.dto.product.ProductDTO;
 import com.antique.service.user.UserDibsService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
