+++ src/main/java/com/antique/dto/product/ProductDTO.java
@@ -1,4 +1,4 @@
-package com.antique.dto;
+package com.antique.dto.product;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
