+++ src/test/java/com/antique/service/ProductServiceTest.java
@@ -4,8 +4,8 @@ import com.antique.TestDataFactory;
 import com.antique.domain.Category;
 import com.antique.domain.Product;
 import com.antique.domain.User;
-import com.antique.dto.ProductDTO;
-import com.antique.dto.ProductInfoDTO;
+import com.antique.dto.product.ProductDTO;
+import com.antique.dto.product.ProductInfoDTO;
 import com.antique.exception.product.ProductNotFoundException;
 import com.antique.repository.ProductRepository;
 import org.junit.jupiter.api.BeforeEach;
