+++ src/test/java/com/antique/service/user/UserDibsServiceTest.java
@@ -4,7 +4,7 @@ import com.antique.domain.Category;
 import com.antique.domain.Dibs;
 import com.antique.domain.Product;
 import com.antique.domain.User;
-import com.antique.dto.ProductDTO;
+import com.antique.dto.product.ProductDTO;
 import com.antique.exception.user.UserNotFoundException;
 import com.antique.repository.DibsRepository;
 import com.antique.repository.UserRepository;
